Cutscenes Added and objects in living room added
All scripts for cutscenes were made. The cutscenes are finished except for the second one in library(need help to activate) and the one in the living room(story not finished). The text for the cutscenes happens as intended. I added all objects in kanbanchi into living room and i added colliders. all i need to change for the objects are to let us walk behing them.